cluster/requirements.txt
flask-login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-APScheduler==1.12.2
gevent==21.8.0
gunicorn==20.1.0
requests==2.26.0
click==8.0.1